Try to use environment variable
It will be easier to integrate into the nextflow pipeline if this uses
and environment variable, as programmatically editing this file would be
tricky. This way when we run the pipeline we can just set the variable
and everything should just work. We should probably avoid hardcoded
paths in general as we will be moving between clusters soon.

Changes are currently untested.