Improve logging in add_ref.pl
For some reason this is currently failing. No clear idea why so I've
added a few things, first it actually checks the result of trying to
get a reference, so it will fail if needed. Secondly, it will fail fast
when fetching data, and warn that it is failing. Finally, it will not
use a reference that was not fetched properly. Hopefully this makes it
easier to trackdown the actual problem.