Tolerate 404's in RNAcentral lookup
The new MD5 based endpoint will return a 404 for cases where the MD5 is
unknown. An error response is assumed to be unusual and a error
condition because, previously the API would return a empty list and it
would not fail. This leads to failing the QC of any sequence not in
RNAcentral.

This change makes a failing response treated as not found. This could
cover up errors in the RNAcentral API and then fail later, I assume but
hopefully this would be caught elsewhere.